requirements.txt
python-dotenv
discord
aiohttp